requirements.txt
flask
flask-cors
tencentcloud-sdk-python
pillow
python-dotenv